refactoring architecture for adhering to SRP and a better overview commit 2 -endpoint fixed and more controllers added - tests still need refactoring
