+++ src/main/java/alignpro/Repository/ProjectRepository.java
@@ -1,12 +1,16 @@
 package alignpro.Repository;
 
 import alignpro.Model.DBConnection;
+import alignpro.Model.Project;
 import jakarta.annotation.PostConstruct;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Repository;
 
 import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
 
 @Repository("PROJECT_REPOSITORY_JDBC")
 @Lazy
@@ -26,5 +30,126 @@ public class ProjectRepository implements IProjectRepository {
     @Override
     public void setConn() {
         this.conn = DBConnection.getConnection(dbURL,dbUsername,dbPassword);
+    }
+
+
+    //************************* Save Method *******************************//
+
+    @Override
+    public void saveProject(String projectName, String startDate,String deadLine, String projectDescription){
+
+        try{
+
+            String sqlString = "INSERT INTO Project (ProjectName, StartDate, Deadline, ProjectDescription) VALUES (?,?,?,?)";
+
+            PreparedStatement stmt = conn.prepareStatement(sqlString);
+            stmt.setString(1, projectName);
+            stmt.setString(2, startDate);
+            stmt.setString(3, deadLine);
+            stmt.setString(4, projectDescription);
+            stmt.executeUpdate();
+
+
+        } catch(SQLException e){
+            throw new RuntimeException(e);
+        }
+    }
+
+    //************************* Get by name Method *******************************//
+
+    @Override
+    public Project getProject(String projectName){
+        Project obj = null;
+
+        try{
+            String sqlString = "SELECT ProjectID, ProjectName, StartDate, Deadline, ProjectDescription FROM Project WHERE ProjectName = ?";
+
+            PreparedStatement stmt = conn.prepareStatement(sqlString);
+            stmt.setString(1,projectName);
+
+            ResultSet rls = stmt.executeQuery();
+            if(rls.next()){
+                obj = new Project();
+                obj.setProjectID(rls.getInt("ProjectID"));
+                obj.setProjectName(rls.getString("ProjectName"));
+                obj.setStartDate(rls.getString("StartDate"));
+                obj.setDeadLine(rls.getString("Deadline"));
+                obj.setProjectDescription(rls.getString("ProjectDescription"));
+            }
+
+        } catch (SQLException e){
+            throw new RuntimeException("Problem getting your project from the DB based on Projectname" + e.getMessage());
+        }
+
+        return obj;
+    }
+
+    //************************* Get by ID Method *******************************//
+
+    @Override
+    public Project getProject(int projectID){
+        Project obj = null;
+
+        try{
+            String sqlString = "SELECT ProjectID, ProjectName, StartDate, Deadline, ProjectDescription FROM Project WHERE ProjectID = ?";
+
+            PreparedStatement stmt = conn.prepareStatement(sqlString);
+            stmt.setInt(1,projectID);
+
+            ResultSet rls = stmt.executeQuery();
+            if(rls.next()){
+                obj = new Project();
+                obj.setProjectID(rls.getInt("ProjectID"));
+                obj.setProjectName(rls.getString("ProjectName"));
+                obj.setStartDate(rls.getString("StartDate"));
+                obj.setDeadLine(rls.getString("Deadline"));
+                obj.setProjectDescription(rls.getString("ProjectDescription"));
+            }
+
+        } catch (SQLException e){
+            throw new RuntimeException("Problem getting your project from the DB based on ProjectID" + e.getMessage());
+        }
+
+        return obj;
+    }
+
+    //************************* Edit Method *******************************//
+
+    @Override
+    public void editProject(Project project, int projectID) {
+        String sqlString = "UPDATE Project SET ProjectName = ?, StartDate = ?, Deadline = ?," +
+                " ProjectDescription = ? WHERE ProjectID = ?";
+
+        try {
+            PreparedStatement stmt = conn.prepareStatement(sqlString);
+            stmt.setString(1, project.getProjectName());
+            stmt.setString(2, project.getStartDateString());
+            stmt.setString(3, project.getDeadlineString());
+            stmt.setString(4, project.getProjectDescription());
+            stmt.setInt(5, projectID);
+
+            stmt.executeUpdate();
+
+        } catch (SQLException e) {
+            throw new RuntimeException("Problem updating your project from the DB based on ProjectID" + e.getMessage());
+        }
+
+    }
+
+    //************************* Delete method *******************************//
+
+    @Override
+    public void deleteProject(int ProjectID){
+        String sqlString = "DELETE FROM Project WHERE ProjectID = ?";
+
+        try{
+
+            PreparedStatement stmt = conn.prepareStatement(sqlString);
+            stmt.setInt(1,ProjectID);
+            stmt.executeUpdate();
+
+        } catch (SQLException e){
+            throw new RuntimeException("Not deleting project from DB" + e.getMessage());
+        }
     }
 }
