+++ src/main/java/alignpro/Service/SubProjectService.java
@@ -1,5 +1,6 @@
 package alignpro.Service;
 
+import alignpro.Model.SubProject;
 import alignpro.Repository.ISubProjectRepository;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.ApplicationContext;
@@ -12,5 +13,30 @@ public class SubProjectService {
 
     public SubProjectService(ApplicationContext context, @Value("${subproject.repository.impl}") String impl) {
         subProjectRepository = (ISubProjectRepository) context.getBean(impl);
+    }
+
+    public void saveSubProject(SubProject obj){
+        subProjectRepository.saveSubProject(obj.getSubProjectName(), obj.getStartDateString(),
+                obj.getEndDateString(), obj.getSubProjectDescription(), obj.getFkProjectID());
+    }
+
+    public void editSubProject(SubProject subProject, int subProjectID){
+        try {
+            subProjectRepository.editSubProject(subProject, subProjectID);
+        }catch (Exception e){
+            throw new RuntimeException("error updating subproject " + subProjectID, e);
+        }
+    }
+
+    public SubProject getSubProject(String subProjectName){
+        return subProjectRepository.getSubProject(subProjectName);
+    }
+
+    public SubProject getSubProject(int subProjectID){
+        return subProjectRepository.getSubProject(subProjectID);
+    }
+
+    public void deleteSubProject(int subProjectID){
+        subProjectRepository.deleteSubProject(subProjectID);
     }
 }
